pyproject.toml
[tool.poetry]
name = "eredesscraper"
version = "0"
description = "Selenium-based web scraper to extract data from E-REDES website and load it into database storage."
authors = ["Ricardo Filipe dos Santos <ricardofilipecdsantos@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/rf-santos/eredes-scraper"
repository = "https://github.com/rf-santos/eredes-scraper"
documentation = "https://github.com/rf-santos/eredes-scraper"
keywords = ["selenium", "web", "scraper", "influxdb", "database", "electricity", "energy", "e-redes"]
include = ["README.md", "LICENSE", "pyproject.toml", "config.yml", "requirements.txt", "tests/**"]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
influxdb-client = "^1.38.0"
openpyxl = "^3.1.2"
pandas = "^2.1.1"
pyyaml = "^6.0.1"
selenium = "^4.14.0"
webdriver-manager = "^4.0.1"
typer = {extras = ["all"], version = "^0.9.0"}

[tool.poetry.group.dev.dependencies]
ipython = "^8.17.2"
jupyter = "^1.0.0"
requests = "^2.31.0"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "semver"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry.scripts]
ers = "eredesscraper.cli:app"

